fix(curriculum): amend challenge wording to ensure passing test
Challenge states:
"Set the position property of your .cat-head element to absolute, then set top and left properties to any pixel value."
Entering a negative value for the left property will prevent test passing.  A negative value is valid for the left property and should be permitted.  It may be preferrable in this case, to simply limit the range of acceptable values in the challenge by including the word "positive".  This would eliminate the need to write tests for "rare" side cases.